Avoid built-in old `test-unit`
In Ruby <= 2.1, `test-unit` was shipped with Ruby itself (unbundling
was done for 2.2 -- see <https://bugs.ruby-lang.org/issues/9711>).

The `test-unit` version shipped with 2.1 breaks some of the tests.
To fix this, I've added the minimum needed version explicitly to the
`gemspec`, as well as added a `gems.rb` for allowing the use of
`bundler` to run the tests and ensure the correct `test-unit` is used.